requirement.txt
pytorch
numpy
pillow